Add modal virtual keyboard for menu text editing on touchscreens
- Implement askModalText() as a shared full-screen modal QWERTY virtual keyboard
- Support touch typing, shifted characters, Space, Del, Left/Right cursor movement, tap-to-reposition cursor, and clipboard paste (Android/web)
- Hook tapNavigation() in menu.cpp to open modal keyboard on MENU_TEXT field taps (e.g. Callsign Title(s), PTT, Comment, Watch List)
- Refactor configs.cpp to use shared askModalText()
- Add askmodal.o to Makefile OBJS
